debugging contrastive loss/ nan

+++ llava/model/llava_arch.py
@@ -324,7 +324,9 @@ class LlavaMetaForCausalLM(ABC):
 
 
     def clip_contrastive_loss(self, input_text_embeds, input_vision_embeds, attention_mask):
-    # Zero out the padded tokens in the embeddings using the attention mask
+    
+        attention_mask = attention_mask.float()
+        # Zero out the padded tokens in the embeddings using the attention mask
         expanded_mask = attention_mask.unsqueeze(-1).expand_as(input_text_embeds)
         # input_text_embeds = input_text_embeds * expanded_mask
